fix: fixed the bug of search input text
When you search something and the text has capital letters, the app will change the searched text to small letters so we wouldn't get an error. But when you go to a song page and go back to the search page, the input text is changed. All letters of input are small. Now this is not going to happen. If all of your searched text's letters are capital letters, they will remain capital and you'll get the songs you want no matter your text's letters are capital or small.