fix: fixed the nothing was found bug
So before when we were searching somthing, if the result was empty the nothing was found text would've been shown to us.
But after that if we were searching another thing and the result wasn't empty, it still would've shown nothing was found text.
So now this bug is fixed.